Added: - CaseDbo and CaseDto separation - function to insert data into database - function to get Response - function to init database - also some code cleanUp
