+++ src/core/Application.kt
@@ -1,5 +1,7 @@
 package com.ilya.shevtsov.casewatcher.core
 
+import core.Server
+
 fun main() {
 
     val server = Server()
