+++ src/model/CaseDataResponse.kt
@@ -1,4 +1,4 @@
-package com.ilya.shevtsov.casewatcher.Model
+package model
 
 import com.google.gson.annotations.SerializedName
 import kotlinx.serialization.Serializable
